Game functionality update

+++ scripts/saving.js
@@ -1,18 +1,20 @@
+let paused = false;
+
 setInterval(() => {
-  localStorage.clear();
-  localStorage.setItem("State", JSON.stringify(grid));
-  localStorage.setItem("Xp", JSON.stringify(xp));
-}, 5000);
+  if (!paused) {
+    localStorage.clear();
+    localStorage.setItem("State", JSON.stringify(grid));
+    localStorage.setItem("Xp", JSON.stringify(xp));
+  }
+}, 500);
 
 function reset() {
+  paused = true;
   const confirmation = confirm(
     "================== Are you sure? ================== \n Note: this will remove all the items currently on conveyor belts"
   );
-  //FIXME: maybe overkill but sometimes doesn't clear the storage
   if (confirmation) {
-    while (localStorage.getItem("State")) {
-      localStorage.clear();
-    }
+    localStorage.clear();
     location.reload();
   }
 }
